NodeConnectorRight, Node IO polling function

+++ include/Widgets/NodeEditor/NodeConnectorLeft.hpp
@@ -7,7 +7,7 @@ class NodeContainer;
 class NodeConnectorLeft : public QWidget
 {
 public:
-    NodeConnectorLeft(NodeContainer* parent, NodeInput* input, int y);
+    NodeConnectorLeft(NodeContainer* parent, NodeInput* input, float y);
 protected:
     NodeInput* input;
     virtual void paintEvent(QPaintEvent*);
